+++ gw/components/interconnect/rtl/data_bus.sv
@@ -1,4 +1,4 @@
-//This is BoxLambda's data bus: a 4-to-1 arbiter connected to a 1-to-19
+//This is BoxLambda's data bus: a 4-to-1 arbiter connected to a 1-to-18
 //mux.
 module data_bus #(
     parameter DATA_WIDTH = 32,  // width of data bus in bits (8, 16, 32, or 64)
@@ -7,13 +7,13 @@ module data_bus #(
     parameter ARB_TYPE_ROUND_ROBIN = 0,  // select round robin arbitration
     parameter ARB_LSB_HIGH_PRIORITY = 1,  // LSB priority selection
     parameter ARB_BLOCK_ACK = 1,  // block ack generation
-    parameter [19*ADDR_WIDTH-1:0] SLAVE_ADDRESSES,
-    parameter [19*ADDR_WIDTH-1:0] SLAVE_ADDR_MASKS
+    parameter [18*ADDR_WIDTH-1:0] SLAVE_ADDRESSES,
+    parameter [18*ADDR_WIDTH-1:0] SLAVE_ADDR_MASKS
 ) (
     input wire clk,
     input wire rst,
     wb_if.slave wbm[4],
-    wb_if.master wbs[19]
+    wb_if.master wbs[18]
 );
 
   import wb_pkg::*;
@@ -38,7 +38,7 @@ module data_bus #(
       .wbs(arbiter_to_mux_if)
   );
 
-  wb_mux_19_wrapper #(
+  wb_mux_18_wrapper #(
       .DATA_WIDTH(DATA_WIDTH),
       .ADDR_WIDTH(ADDR_WIDTH),
       .SELECT_WIDTH(SELECT_WIDTH),